A new and more straightforward method to finding a palindrome
Instead of comparing each character of the original string to that of its reverse through the entire string, I decided to compare the first letter and the last letter of the string. If the start and end character equal eachother it is a palindrome and as long as start is less than end the string is considered to be a palindrome. This allows for both even and odd length strings to be correctly determined